requirements.txt
numpy==1.22.1
onnx==1.10.2
onnxoptimizer==0.2.6
onnxruntime==1.10.0
Pillow==9.0.0
